+++ lib/arel/extensions.rb
@@ -21,6 +21,7 @@ require 'arel/extensions/range_function'
 require 'arel/extensions/with_ordinality'
 require 'arel/extensions/table'
 require 'arel/extensions/row'
+require 'arel/extensions/ordering'
 
 require 'arel/extensions/functions'
 require 'arel/extensions/nodes'
