+++ lib/arel/extensions/ordering.rb
@@ -1,0 +1,36 @@
+# rubocop:disable Naming/MethodName
+# rubocop:disable Naming/UncommunicativeMethodParamName
+
+module Arel
+  module Nodes
+    Arel::Nodes::Ordering.class_eval do
+      # Postgres: https://www.postgresql.org/docs/9.4/queries-order.html
+      attr_accessor :nulls
+
+      def initialize(expr, nulls = 0)
+        super(expr)
+
+        @nulls = nulls
+      end
+    end
+  end
+
+  module Visitors
+    class ToSql
+      alias old_visit_Arel_Nodes_Ascending visit_Arel_Nodes_Ascending
+      def visit_Arel_Nodes_Ascending(o, collector)
+        old_visit_Arel_Nodes_Ascending(o, collector)
+        apply_ordering_nulls(o, collector)
+      end
+
+      alias old_visit_Arel_Nodes_Descending visit_Arel_Nodes_Descending
+      def visit_Arel_Nodes_Descending(o, collector)
+        old_visit_Arel_Nodes_Descending(o, collector)
+        apply_ordering_nulls(o, collector)
+      end
+    end
+  end
+end
+
+# rubocop:enable Naming/MethodName
+# rubocop:enable Naming/UncommunicativeMethodParamName
